+++ lib/procrastinator/rspec/matchers.rb
@@ -16,8 +16,15 @@ RSpec::Matchers.define :have_task do |expected_task|
       expected = a_hash_including(expected_task)
 
       actual_tasks = task_store.read.collect do |task|
-         task_hash        = task.to_h
-         task_hash[:data] = JSON.parse(task_hash[:data], symbolize_names: true) unless task_hash[:data].empty?
+         task_hash = task.to_h
+         unless task_hash[:data].nil? || task_hash[:data].empty?
+            task_hash[:data] = JSON.parse(task_hash[:data], symbolize_names: true)
+         end
+         task_hash[:queue] = task_hash[:queue].to_sym if task_hash[:queue]
+         Procrastinator::Task::TIME_FIELDS.each do |time_field|
+            task_hash[time_field] = Time.at(task_hash[time_field].to_i) if task_hash[time_field]&.respond_to?(:to_i)
+         end
+
          task_hash
       end